% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{buildDVImodel}
\alias{buildDVImodel}
\title{Create an input object}
\usage{
buildDVImodel(date, day_len, temp, sowing, heading, critical, acc)
}
\arguments{
\item{date}{a character vector to indicate the dates on which day lengths
and temperature were measured. The length should match the lengths of
\code{day_len} and \code{temp}.}

\item{day_len}{a numeric vector to indicate the dates on which day lengths
and temperature were measured. The length should match the lengths of \code{date}
and \code{temp}.}

\item{temp}{a numeric vector to indicate the dates on which day lengths and
temperature were measured. The length should match the lengths of \code{date}
and \code{day_len}.}

\item{sowing}{a numeric value to indicate the sowing date.}

\item{heading}{a numeric value to indicate the heading date.}

\item{critical}{a numeric value to indicate the critical day length.}

\item{acc}{a character indicates sample ID.}
}
\value{
A DviModel object that can be specified to the \code{model} argument of
\code{\link[=estDVIparam]{estDVIparam()}}
}
\description{
Create an input object for the parameter estimation of the DVI model using
a genetic algorithm.
}
\details{
Dates should be given in the 'YYYY-MM-DD' format.
}
