% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{dviFixedParam}
\alias{dviFixedParam}
\title{Set fixed parameters of DVI models}
\usage{
dviFixedParam(tb = 8, to = 30, tc = 42, pb = 0, po = 10, pc = 24)
}
\arguments{
\item{tb}{a numeric value to specify the base temperature.}

\item{to}{a numeric value to specify the optimal temperature.}

\item{tc}{a numeric value to specify the ceiling temperature.}

\item{pb}{a numeric value to specify the base photoperiod.}

\item{po}{a numeric value to specify the optimal photoperiod.}

\item{pc}{a numeric value to specify the ceiling photoperiod.}
}
\value{
A DviParam object that can be specified to the \code{fixed_param}
argument of \code{\link[=estDVIparam]{estDVIparam()}}
}
\description{
Create a parameter object to specify fixed parameters of a DVI model.
}
