% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{estDVIparam}
\alias{estDVIparam}
\title{DVI model parameter estimation}
\usage{
estDVIparam(object, fixed_param = dviFixedParam(), ga_param = gaParam())
}
\arguments{
\item{object}{a DviModel object created by \code{\link[=buildDVImodel]{buildDVImodel()}}.}

\item{fixed_param}{a DviParam object created by \code{\link[=dviFixedParam]{dviFixedParam()}}.}

\item{ga_parma}{a GaParam object created by \code{\link[=gaParam]{gaParam()}}.}
}
\value{
A DviEst object.
}
\description{
Estimate the parameters of the DVI model using a genetic algorithm.
}
\details{
Dates should be given in the 'YYYY-MM-DD' format.
}
\examples{
fixed_param <- dviFixedParam()
ga_param <- gaParam()
date <- seq.Date(from = as.Date("2023-06-01"),
                 to = as.Date("2023-08-31"),
                 by = "day")
day_len <- c(seq(12, 13.5, length.out = 30),
             seq(13.5, 12, length.out = length(date) - 30))
temp = c(sample(x = 25:30, size = 30, replace = TRUE),
         sample(x = 25:35, size = 31, replace = TRUE),
         sample(x = 30:35, size = 31, replace = TRUE))
dvi_model <- buildDVImodel(date = date,
                           day_len = day_len,
                           temp = temp,
                           sowing = "2023-06-05",
                           heading = "2023-08-20",
                           critical = 13,
                           acc = "test")

dvi_est <- estDVIparam(object = dvi_model,
                       fixed_param = fixed_param,
                       ga_param = ga_param)

}
\seealso{
\code{\link[=predict.DviEst]{predict.DviEst()}}, \code{\link[=summary.DviEst]{summary.DviEst()}} and \link{plotDviEst}
}
