% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{gaParam}
\alias{gaParam}
\title{Set genetic algorithm parameters}
\usage{
gaParam(
  lower = c(0, 0, 30),
  upper = c(40, 40, 150),
  maxiter = 100,
  popSize = 50,
  parallel = 1
)
}
\arguments{
\item{lower}{a vector of length equal to the decision variables providing the
lower bounds of the search space in case of real-valued or permutation
encoded optimizations.}

\item{upper}{a vector of length equal to the decision variables providing the
upper bounds of the search space in case of real-valued or permutation
encoded optimizations.}

\item{maxiter}{the maximum number of iterations to run before the GA search
is halted.}

\item{popSize}{the population size.}

\item{parallel}{An optional argument which allows to specify if the Genetic
Algorithm should be run sequentially or in parallel. See \code{\link[GA:ga]{GA::ga()}} for more
details.}
}
\value{
A GaParam object that can be specified to the \code{ga_param} argument of
\code{\link[=estDVIparam]{estDVIparam()}}
}
\description{
Create a parameter object to control the parameter estimation by a genetic
algorithm.
}
\seealso{
GA::ga()
}
