% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{getCor}
\alias{getCor}
\title{Get the correlation coefficient

Extracted the correlation coefficient from the DviEst object.}
\usage{
getCor(object)
}
\arguments{
\item{object}{a DviEstl object created by \code{\link[=estDVIparam]{estDVIparam()}}.}
}
\value{
a numeric value of the correlation coefficient.
}
\description{
Get the correlation coefficient

Extracted the correlation coefficient from the DviEst object.
}
