% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{getLM}
\alias{getLM}
\title{Get the linear model}
\usage{
getLM(object)
}
\arguments{
\item{object}{a DviEstl object created by \code{\link[=estDVIparam]{estDVIparam()}}.}
}
\value{
a lm class object.
}
\description{
Extracted the linear model that represents the relationship between
the observed and predicted heading dates.
}
