% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{getPred}
\alias{getPred}
\title{Get predicted heading dates}
\usage{
getPred(object)
}
\arguments{
\item{object}{a DviEstl object created by \code{\link[=estDVIparam]{estDVIparam()}}.}
}
\value{
a data.frame extracted from the prediction slot of the DviEst object.
}
\description{
Extracted predicted heading dates and days-to-heading data
from the DviEst object.
}
