% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{plotDviEst}
\alias{plotDviEst}
\title{Plot scatter plot of correlation}
\usage{
plotDviEst(object)
}
\arguments{
\item{object}{a DviEst object created by \code{\link[=estDVIparam]{estDVIparam()}}.}
}
\value{
A ggplot2 object.
}
\description{
Plot a scatter plot to visualize correlation between the observed heading
dates and predicated heading dates.
}
\examples{
fixed_param <- dviFixedParam()
ga_param <- gaParam()
date <- seq.Date(from = as.Date("2023-06-01"),
                 to = as.Date("2023-08-31"),
                 by = "day")
day_len <- c(seq(12, 13.5, length.out = 30),
             seq(13.5, 12, length.out = length(date) - 30))
temp = c(sample(x = 25:30, size = 30, replace = TRUE), 
         sample(x = 25:35, size = 31, replace = TRUE),
         sample(x = 30:35, size = 31, replace = TRUE))
dvi_model <- buildDVImodel(date = date,
                           day_len = day_len,
                           temp = temp,
                           sowing = "2023-06-05",
                           heading = "2023-08-20", 
                           critical = 13,
                           acc = "test")
                           
dvi_est <- estDVIparam(object = dvi_model, 
                       fixed_param = fixed_param,
                       ga_param = ga_param)

dvi_est <- predict(object = dvi_est)

p <- plotDviEst(object = dvi_est)

print(p)

}
\seealso{
\code{\link[=predict.DviEst]{predict.DviEst()}}, \code{\link[=summary.DviEst]{summary.DviEst()}} and \link{plotDviEst}
}
