% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{predict.DviEst}
\alias{predict.DviEst}
\title{DVI model parameter estimation}
\usage{
\method{predict}{DviEst}(object, new_model = NULL)
}
\arguments{
\item{object}{a DviEst object created by \code{\link[=estDVIparam]{estDVIparam()}}.}

\item{fixed_param}{a DviParam object created by \code{\link[=dviFixedParam]{dviFixedParam()}}.}

\item{ga_parma}{a GaParam object created by \code{\link[=gaParam]{gaParam()}}.}

\item{new_mode}{a DviModel object created by \code{\link[=buildDVImodel]{buildDVImodel()}}.}
}
\value{
A DviEst object.

A DviEst object.
}
\description{
Estimate the parameters of the DVI model using a genetic algorithm.
}
\details{
Dates should be given in the 'YYYY-MM-DD' format.

The predicated heading date (HD_Pred) and the days to heading
(DTH_Pred) will be returned as a data.frame in the prediction slot of the
output DviEst object. If the observed heading dates are available in the
input DviEst object, these values also returned at the DTH_Obs column of
the data.frame in the prediction slot. The predicted heading dates will be
compared with the observed heading dates recorded in the given DviEst object,
and the correlation coefficient between these will be returned in the
correlation slot of the DviEst object.
If a DviModle object was specified to \code{new_model}, only the predicated
heading date (HD_Pred) and days to heading (DTH_Pred) will be returned in the
prediction slot.
}
\examples{
fixed_param <- dviFixedParam()
ga_param <- gaParam()
date <- seq.Date(from = as.Date("2023-06-01"),
                 to = as.Date("2023-08-31"),
                 by = "day")
day_len <- c(seq(12, 13.5, length.out = 30),
             seq(13.5, 12, length.out = length(date) - 30))
temp = c(sample(x = 25:30, size = 30, replace = TRUE),
         sample(x = 25:35, size = 31, replace = TRUE),
         sample(x = 30:35, size = 31, replace = TRUE))
dvi_model <- buildDVImodel(date = date,
                           day_len = day_len,
                           temp = temp,
                           sowing = "2023-06-05",
                           heading = "2023-08-20",
                           critical = 13,
                           acc = "test")

dvi_est <- estDVIparam(object = dvi_model,
                       fixed_param = fixed_param,
                       ga_param = ga_param)

Heading date prediction

Predict heading dates based on sowing dates and the DVI model with the
estimated parameters.

fixed_param <- dviFixedParam()
ga_param <- gaParam()
date <- seq.Date(from = as.Date("2023-06-01"),
                 to = as.Date("2023-08-31"),
                 by = "day")
day_len <- c(seq(12, 13.5, length.out = 30),
             seq(13.5, 12, length.out = length(date) - 30))
temp = c(sample(x = 25:30, size = 30, replace = TRUE),
         sample(x = 25:35, size = 31, replace = TRUE),
         sample(x = 30:35, size = 31, replace = TRUE))
dvi_model <- buildDVImodel(date = date,
                           day_len = day_len,
                           temp = temp,
                           sowing = "2023-06-05",
                           heading = "2023-08-20",
                           critical = 13,
                           acc = "test")

dvi_est <- estDVIparam(object = dvi_model,
                       fixed_param = fixed_param,
                       ga_param = ga_param)

dvi_est <- predict(object = dvi_est)

}
\seealso{
\code{\link[=predict.DviEst]{predict.DviEst()}}, \code{\link[=summary.DviEst]{summary.DviEst()}} and \link{plotDviEst}

\code{\link[=predict.DviEst]{predict.DviEst()}}, \code{\link[=summary.DviEst]{summary.DviEst()}} and \link{plotDviEst}
}
