% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{setDviParam}
\alias{setDviParam}
\title{Set parameters of a DVI model}
\usage{
setDviParam(object, fixed_param, alpha, beta, g)
}
\arguments{
\item{object}{a DviModel object created by \code{\link[=buildDVImodel]{buildDVImodel()}}.}

\item{fixed_param}{a DviParam object created by \code{\link[=dviFixedParam]{dviFixedParam()}}.}

\item{alpha}{a numeric value of the temperature-sensitivity coefficient.}

\item{beta}{a numeric value of the photoperiod-sensitivity coefficient.}

\item{g}{a numeric value of the earliness of flowering under optimal
photoperiod and temperature.}
}
\value{
a DviEst object.
}
\description{
Set arbitrary values as estimated parameters of a DVI model.
}
