% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{summary.DviEst}
\alias{summary.DviEst}
\title{Show summary of a DviEst object}
\usage{
\method{summary}{DviEst}(x)
}
\arguments{
\item{x}{a DviEst object created by \code{\link[=estDVIparam]{estDVIparam()}}.}
}
\value{
Invisibly return a data.frame of the summary.
}
\description{
Show summary of a DviEst object
}
