% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_functions.R
\name{summary.DviModel}
\alias{summary.DviModel}
\title{Show summary of a DviModel object}
\usage{
\method{summary}{DviModel}(x)
}
\arguments{
\item{x}{a DviModel object created by \code{\link[=buildDVImodel]{buildDVImodel()}}.}
}
\value{
Invisibly return the input object itself.
}
\description{
Show summary of a DviModel object
}
